+++ registry/src/lit/ui/block-handle/block-handle.ts
@@ -40,3 +40,14 @@ export class LitBlockHandle extends LitElement {
     </prosekit-block-handle-root>`
   }
 }
+
+export function registryLitEditorBlockHandle() {
+  if (customElements.get('lit-editor-block-handle')) return
+  customElements.define('lit-editor-block-handle', LitBlockHandle)
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'lit-editor-block-handle': LitBlockHandle
+  }
+}
