+++ registry/src/lit/ui/block-handle/index.ts
@@ -1,12 +1,1 @@
-import { LitBlockHandle } from './block-handle'
-
-export function registryLitEditorBlockHandle() {
-  if (customElements.get('lit-editor-block-handle')) return
-  customElements.define('lit-editor-block-handle', LitBlockHandle)
-}
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'lit-editor-block-handle': LitBlockHandle
-  }
-}
+export { registryLitEditorBlockHandle } from './block-handle'
