+++ registry/src/lit/ui/button/button.ts
@@ -58,3 +58,14 @@ export class LitButton extends LitElement {
     `
   }
 }
+
+export function registryLitEditorButton() {
+  if (customElements.get('lit-editor-button')) return
+  customElements.define('lit-editor-button', LitButton)
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'lit-editor-button': LitButton
+  }
+}
