+++ registry/src/lit/ui/button/index.ts
@@ -1,12 +1,1 @@
-import { LitButton } from './button'
-
-export function registryLitEditorButton() {
-  if (customElements.get('lit-editor-button')) return
-  customElements.define('lit-editor-button', LitButton)
-}
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'lit-editor-button': LitButton
-  }
-}
+export { registryLitEditorButton } from './button'
