+++ registry/src/lit/ui/drop-indicator/drop-indicator.ts
@@ -27,3 +27,14 @@ export class LitDropIndicator extends LitElement {
     ></prosekit-drop-indicator>`
   }
 }
+
+export function registryLitEditorDropIndicator() {
+  if (customElements.get('lit-editor-drop-indicator')) return
+  customElements.define('lit-editor-drop-indicator', LitDropIndicator)
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'lit-editor-drop-indicator': LitDropIndicator
+  }
+}
