+++ registry/src/lit/ui/drop-indicator/index.ts
@@ -1,12 +1,1 @@
-import { LitDropIndicator } from './drop-indicator'
-
-export function registryLitEditorDropIndicator() {
-  if (customElements.get('lit-editor-drop-indicator')) return
-  customElements.define('lit-editor-drop-indicator', LitDropIndicator)
-}
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'lit-editor-drop-indicator': LitDropIndicator
-  }
-}
+export { registryLitEditorDropIndicator } from './drop-indicator'
