+++ registry/src/lit/ui/image-upload-popover/index.ts
@@ -1,1 +1,1 @@
-export {registryLitEditorImageUploadPopover}from './image-upload-popover'
+export { registryLitEditorImageUploadPopover } from './image-upload-popover'
