+++ registry/src/lit/ui/table-handle/index.ts
@@ -1,12 +1,1 @@
-import { LitTableHandle } from './table-handle'
-
-export function registryLitEditorTableHandle() {
-  if (customElements.get('lit-editor-table-handle')) return
-  customElements.define('lit-editor-table-handle', LitTableHandle)
-}
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'lit-editor-table-handle': LitTableHandle
-  }
-}
+export { registryLitEditorTableHandle } from './table-handle'
